project_DjangoFace/requirements.txt
asgiref==3.2.7
astropy==4.0.1
click==7.1.1
Django==3.0.5
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-user-accounts==3.0.0
dlib==19.8.1
face-recognition==1.3.0
face-recognition-models==0.3.0
joblib==0.14.1
mysqlclient==1.4.6
numpy==1.18.2
Pillow==7.0.0
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
sqlparse==0.3.1
